Day 5 part 1 works, ugly

+++ src/day05.rs
@@ -17,15 +17,15 @@ enum ParameterMode {
 }
 
 enum Instruction {
-    Add((usize, ParameterMode), (usize, ParameterMode), usize),
-    Mul((usize, ParameterMode), (usize, ParameterMode), usize),
-    Input(usize),
-    Output((usize, ParameterMode)),
+    Add((isize, ParameterMode), (isize, ParameterMode), isize),
+    Mul((isize, ParameterMode), (isize, ParameterMode), isize),
+    Input(isize),
+    Output((isize, ParameterMode)),
     Terminate,
 }
 
 impl Instruction {
-    fn size(&self) -> usize {
+    fn size(&self) -> isize {
         match self {
             Instruction::Add(_,_,_) => 4,
             Instruction::Mul(_,_,_) => 4,
@@ -36,7 +36,7 @@ impl Instruction {
     }
 }
 
-fn parse_parameter_mode(x: usize) -> Result<ParameterMode, &'static str> {
+fn parse_parameter_mode(x: isize) -> Result<ParameterMode, &'static str> {
     match x {
         0 => Ok(ParameterMode::Positional),
         1 => Ok(ParameterMode::Immediate),
@@ -44,7 +44,7 @@ fn parse_parameter_mode(x: usize) -> Result<ParameterMode, &'static str> {
     }
 }
 
-fn parse_instruction(program: &[usize], ip: usize) -> Result<Instruction, &'static str> {
+fn parse_instruction(program: &[isize], ip: usize) -> Result<Instruction, &'static str> {
     let val0 = *program.get(ip).ok_or("out of bounds")?;
     let opcode = val0 % 100;
     if opcode == 99 {
@@ -78,36 +78,36 @@ fn parse_instruction(program: &[usize], ip: usize) -> Result<Instruction, &'stat
     Err("invalid opcode")
 }
 
-fn read_parameter((a, pm): (usize, ParameterMode), program: &[usize]) -> Result<usize, &'static str> {
+fn read_parameter((a, pm): (isize, ParameterMode), program: &[isize]) -> Result<isize, &'static str> {
     match pm {
         ParameterMode::Immediate => Ok(a),
         ParameterMode::Positional => {
-            let x1 = program.get(a).ok_or("out of bounds")?;
+            let x1 = program.get(a as usize).ok_or("out of bounds")?;
             Ok(*x1)
         }
     }
 }
 
-fn execute_instruction(program: &mut Vec<usize>, ip: usize) -> Result<Option<usize>, &'static str> {
-    let inst = parse_instruction(&program, ip)?;
+fn execute_instruction(program: &mut Vec<isize>, ip: isize) -> Result<Option<isize>, &'static str> {
+    let inst = parse_instruction(&program, ip as usize)?;
     match inst {
         Instruction::Terminate => Ok(None),
         Instruction::Add(i1, i2, i3) => {
             let x1 = read_parameter(i1, program)?;
             let x2 = read_parameter(i2, program)?;
-            program[i3] = x1 + x2;
+            program[i3 as usize] = x1 + x2;
             Ok(Some(ip + inst.size()))
         }
         Instruction::Mul(i1, i2, i3) => {
             let x1 = read_parameter(i1, program)?;
             let x2 = read_parameter(i2, program)?;
-            program[i3] = x1 * x2;
+            program[i3 as usize] = x1 * x2;
             Ok(Some(ip + inst.size()))
         }
         Instruction::Input(i1) => {
             let hc_in = 1;
             println!("input asked, giving {}", hc_in);
-            program[i1] = hc_in;
+            program[i1 as usize] = hc_in;
             Ok(Some(ip + inst.size()))
         }
         Instruction::Output(i1) => {
@@ -118,8 +118,8 @@ fn execute_instruction(program: &mut Vec<usize>, ip: usize) -> Result<Option<usi
     }
 }
 
-fn execute_until_termination(program: &mut Vec<usize>) -> Result<usize, &'static str> {
-    let mut ip: usize = 0;
+fn execute_until_termination(program: &mut Vec<isize>) -> Result<isize, &'static str> {
+    let mut ip: isize = 0;
     loop {
         let new_ip = execute_instruction(program, ip)?;
         match new_ip {
@@ -132,34 +132,34 @@ fn execute_until_termination(program: &mut Vec<usize>) -> Result<usize, &'static
     Ok(program[0])
 }
 
-pub fn part1() -> io::Result<usize> {
+pub fn part1() -> io::Result<isize> {
     let mut x = read_input()?;
     let result = execute_until_termination(&mut x).unwrap();
     Ok(result)
 }
 
-fn execute_with_input(program: &[usize], x1: usize, x2: usize) -> Result<usize, &'static str> {
-    let mut prog_copy: Vec<usize> = program.to_vec();
+fn execute_with_input(program: &[isize], x1: isize, x2: isize) -> Result<isize, &'static str> {
+    let mut prog_copy: Vec<isize> = program.to_vec();
     prog_copy[1] = x1;
     prog_copy[2] = x2;
     execute_until_termination(&mut prog_copy)
 }
 
 use clap::{value_t_or_exit, App, Arg};
-pub fn part2_manual() -> io::Result<usize> {
+pub fn part2_manual() -> io::Result<isize> {
     let matches = App::new("run computer")
         .arg(Arg::with_name("x0"))
         .arg(Arg::with_name("x1"))
         .get_matches();
-    let x0 = value_t_or_exit!(matches, "x0", usize);
-    let x1 = value_t_or_exit!(matches, "x1", usize);
+    let x0 = value_t_or_exit!(matches, "x0", isize);
+    let x1 = value_t_or_exit!(matches, "x1", isize);
 
     let x = read_input()?;
     let result = execute_with_input(&x, x0, x1).unwrap();
     Ok(result)
 }
 
-pub fn part2() -> io::Result<(usize, usize)> {
+pub fn part2() -> io::Result<(isize, isize)> {
     let x = read_input()?;
     for x1 in 0..100 {
         for x2 in 0..100 {
@@ -174,7 +174,7 @@ pub fn part2() -> io::Result<(usize, usize)> {
 
 #[test]
 fn test_execute_instruction() {
-    let mut program: Vec<usize> = vec![2, 4, 4, 5, 99, 0];
+    let mut program: Vec<isize> = vec![2, 4, 4, 5, 99, 0];
     execute_instruction(&mut program, 0).unwrap();
     assert_eq!(program[5], 9801);
 }
